By default tee_tty will run in background. Fix bug in Termios when setting NL,CR options from SerialOptions. SerialPort::createPtyLink: don't re-create symlinks if they're OK.

+++ src/nidas/util/SerialPort.cc
@@ -333,6 +333,7 @@ int SerialPort::createPtyLink(const std::string& link) throw(IOException)
     if (grantpt(fd) < 0) throw IOException(ptmx,"grantpt",errno);
     if (unlockpt(fd) < 0) throw IOException(ptmx,"unlockpt",errno);
 
+    bool dolink = true;
     struct stat linkstat;
     if (lstat(link.c_str(),&linkstat) < 0) {
         if (errno != ENOENT)
@@ -340,17 +341,28 @@ int SerialPort::createPtyLink(const std::string& link) throw(IOException)
     }
     else {
         if (S_ISLNK(linkstat.st_mode)) {
-	    cerr << link << " is a symbolic link, deleting" << endl;
-	    if (unlink(link.c_str()) < 0)
-		throw IOException(link,"unlink",errno);
+	    char linkdest[PATH_MAX];
+	    int ld = readlink(link.c_str(),linkdest,PATH_MAX-1);
+	    if (ld < 0)
+		throw IOException(link,"readlink",errno);
+	    linkdest[ld] = 0;
+	    if (strcmp(slave,linkdest)) {
+		cerr << "Deleting " << link << " (a symbolic link to " << linkdest << ")" << endl;
+		if (unlink(link.c_str()) < 0)
+		    throw IOException(link,"unlink",errno);
+	    }
+	    else dolink = false;
 	}
 	else
 	    throw IOException(link,
 	    	"exists and is not a symbolic link","");
 
     }
-    if (symlink(slave,link.c_str()) < 0)
-	throw IOException(link,"symlink",errno);
+    if (dolink) {
+	cerr << "Linking " << slave << " to " << link << endl;
+        if (symlink(slave,link.c_str()) < 0)
+	    throw IOException(link,"symlink",errno);
+    }
     return fd;
 }
